Find maximum number GFG
Given a number N, write a program to find a maximum number that can be formed using all of the digits of this number.
Note: The given number can be very large, so the input is taken as a String.

 

Example 1:

Input:
N = "38293367"
Output:
98763332 
Explanation:
98763332 is the largest number that
can be formed by rearranging the
digits of N.
Example 2:

Input:
N = "1203465"
Output:
6543210
Explanation:
6543210 is the largest number that
can be formed by rearranging the
digits of N.
 

Your Task:
You don't need to read input or print anything. Your task is to complete the function findMax() which takes a String N as input and returns the answer.